requirements.txt
absl-py==1.4.0
appdirs==1.4.4
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
cloudpickle==2.2.1
colorama==0.4.6
contourpy==1.1.0
cycler==0.11.0
docker-pycreds==0.4.0
fonttools==4.42.1
gitdb==4.0.10
GitPython==3.1.36
google-auth==2.23.0
google-auth-oauthlib==0.4.6
grpcio==1.58.0
gym==0.23.1
gym-notices==0.0.8
idna==3.4
kiwisolver==1.4.5
Markdown==3.4.4
MarkupSafe==2.1.3
matplotlib==3.7.3
numpy==1.25.2
oauthlib==3.2.2
packaging==23.1
pandas==2.1.0
pathtools==0.1.2
Pillow==10.0.0
protobuf==3.20.3
psutil==5.9.5
pyasn1==0.5.0
pyasn1-modules==0.3.0
pygame==2.1.0
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
sentry-sdk==1.30.0
setproctitle==1.3.2
six==1.16.0
smmap==5.0.0
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
torch==1.12.1
typing_extensions==4.7.1
tzdata==2023.3
urllib3==1.26.16
wandb==0.15.10
Werkzeug==2.3.7
